# lib/api-zod/src/generated/types/pnlPoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trading Journal API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PnlPointResult } from './pnlPointResult';

export interface PnlPoint {
  tradeNumber: number;
  pnl: number;
  result: PnlPointResult;
  ticker: string;
  date: string;
}