# lib/api-zod/src/generated/types/trade.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trading Journal API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TradeResult } from './tradeResult';
import type { TradeSide } from './tradeSide';

export interface Trade {
  id: number;
  date: string;
  ticker: string;
  side: TradeSide;
  /** @nullable */
  strategy?: string | null;
  entryPrice: number;
  exitPrice: number;
  stopLoss: number;
  shares: number;
  /** @nullable */
  notes?: string | null;
  netPnl: number;
  rr: number;
  result: TradeResult;
  createdAt: string;
}