# lib/api-zod/src/generated/types/tradeInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trading Journal API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TradeInputSide } from './tradeInputSide';

export interface TradeInput {
  date: string;
  ticker: string;
  side: TradeInputSide;
  strategy?: string;
  entryPrice: number;
  exitPrice: number;
  stopLoss: number;
  shares: number;
  notes?: string;
}