# lib/api-zod/src/generated/types/tradeStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trading Journal API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PnlPoint } from './pnlPoint';

export interface TradeStats {
  totalTrades: number;
  winRate: number;
  netPnl: number;
  profitFactor: number;
  wins: number;
  losses: number;
  pnlByTrade: PnlPoint[];
}